requirements.txt
Django==1.4
amqplib==1.0.2
anyjson==0.3.3
billiard==2.7.3.10
celery==3.0.3
distribute==0.6.27
dj-database-url==0.2.1
django-celery==3.0.1
django-mailchimp==0.1.30
facebook-sdk==0.3.2
gunicorn==0.14.5
kombu==2.3.0
psycopg2==2.4.5
python-dateutil==1.5
python-memcached==1.48
requests==0.14.2
wsgiref==0.1.2
pylibmc==1.2.2
django-pylibmc-sasl==0.2.4